Add support for admin users

+++ app/Http/Controllers/CRUDController.php
@@ -4,11 +4,21 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Product;
+use App\Models\User;
 
 class CRUDController extends Controller
 {
     public function index()
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         $products = Product::latest()->paginate(5);
 
         return view('products.index', compact('products'))
@@ -17,11 +27,29 @@ class CRUDController extends Controller
 
     public function create()
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         return view('products.create');
     }
 
     public function store(Request $request)
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -37,16 +65,43 @@ class CRUDController extends Controller
 
     public function show(Product $product)
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         return view('products.show', compact('product'));
     }
 
     public function edit(Product $product)
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         return view('products.edit', compact('product'));
     }
 
     public function update(Request $request, Product $product)
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -62,6 +117,15 @@ class CRUDController extends Controller
 
     public function destroy(Product $product)
     {
+        $user = User::find(auth()->id());
+        if (!$user) {
+            return redirect()->route('products:index');
+        }
+        else if (!$user->is_admin)
+        {
+            return redirect()->route('products:index');
+        }
+
         $product->delete();
 
         return redirect()->route('products.index')
